+++ openseries/_common_model.py
@@ -18,7 +18,15 @@ if TYPE_CHECKING:
 from openpyxl.utils.dataframe import dataframe_to_rows
 from openpyxl.workbook.workbook import Workbook
 from openpyxl.worksheet.worksheet import Worksheet
-from pandas import DataFrame, DatetimeIndex, Index, MultiIndex, Series, date_range
+from pandas import (
+    DataFrame,
+    DatetimeIndex,
+    Index,
+    MultiIndex,
+    Series,
+    date_range,
+    to_datetime,
+)
 from pandas.tseries.offsets import CustomBusinessDay
 from plotly.graph_objs import Figure  # type: ignore[import-untyped,unused-ignore]
 from plotly.io import to_html  # type: ignore[import-untyped,unused-ignore]
@@ -351,9 +359,7 @@ class _CommonModel(BaseModel):
         """
         wmdf = self.tsdf.copy()
         wmdf.index = DatetimeIndex(wmdf.index)
-        result = (
-            wmdf.resample("BME").last().pct_change(fill_method=cast(str, None)).min()
-        )
+        result = wmdf.resample("BME").last().pct_change().min()
 
         if self.tsdf.shape[1] == 1:
             return float(result.iloc[0])
@@ -521,8 +527,8 @@ class _CommonModel(BaseModel):
             An OpenFrame object
 
         """
-        startyear = DatetimeIndex(self.tsdf.index)[0].year
-        endyear = DatetimeIndex(self.tsdf.index)[-1].year
+        startyear = to_datetime(self.tsdf.index[0]).year
+        endyear = to_datetime(self.tsdf.index[-1]).year
         calendar = holiday_calendar(
             startyear=startyear,
             endyear=endyear,
@@ -1021,9 +1027,7 @@ class _CommonModel(BaseModel):
             time_factor = how_many / fraction
 
         result = (
-            self.tsdf.loc[cast(int, earlier) : cast(int, later)]
-            .pct_change(fill_method=cast(str, None))
-            .mean()
+            self.tsdf.loc[cast(int, earlier) : cast(int, later)].pct_change().mean()
             * time_factor
         )
 
@@ -1081,9 +1085,7 @@ class _CommonModel(BaseModel):
             time_factor = how_many / fraction
 
         data = self.tsdf.loc[cast(int, earlier) : cast(int, later)]
-        result = (
-            data.pct_change(fill_method=cast(str, None)).std().mul(sqrt(time_factor))
-        )
+        result = data.pct_change().std().mul(sqrt(time_factor))
 
         if self.tsdf.shape[1] == 1:
             return float(cast(SupportsFloat, result.iloc[0]))
@@ -1279,22 +1281,20 @@ class _CommonModel(BaseModel):
         if drift_adjust:
             imp_vol = (-sqrt(time_factor) / norm.ppf(level)) * (
                 self.tsdf.loc[cast(int, earlier) : cast(int, later)]
-                .pct_change(fill_method=cast(str, None))
+                .pct_change()
                 .quantile(1 - level, interpolation=interpolation)
                 - self.tsdf.loc[cast(int, earlier) : cast(int, later)]
-                .pct_change(fill_method=cast(str, None))
+                .pct_change()
                 .sum()
                 / len(
-                    self.tsdf.loc[cast(int, earlier) : cast(int, later)].pct_change(
-                        fill_method=cast(str, None),
-                    ),
+                    self.tsdf.loc[cast(int, earlier) : cast(int, later)].pct_change(),
                 )
             )
         else:
             imp_vol = (
                 -sqrt(time_factor)
                 * self.tsdf.loc[cast(int, earlier) : cast(int, later)]
-                .pct_change(fill_method=cast(str, None))
+                .pct_change()
                 .quantile(1 - level, interpolation=interpolation)
                 / norm.ppf(level)
             )
@@ -1357,14 +1357,14 @@ class _CommonModel(BaseModel):
         cvar_df = self.tsdf.loc[cast(int, earlier) : cast(int, later)].copy(deep=True)
         result = [
             cvar_df.loc[:, x]  # type: ignore[call-overload,index]
-            .pct_change(fill_method=cast(str, None))
+            .pct_change()
             .sort_values()
             .iloc[
                 : int(
                     ceil(
                         (1 - level)
                         * cvar_df.loc[:, x]  # type: ignore[index]
-                        .pct_change(fill_method=cast(str, None))
+                        .pct_change()
                         .count(),
                     ),
                 )
@@ -1424,7 +1424,7 @@ class _CommonModel(BaseModel):
         )
         how_many = (
             self.tsdf.loc[cast(int, earlier) : cast(int, later)]
-            .pct_change(fill_method=cast(str, None))
+            .pct_change()
             .count(numeric_only=True)
         )
         if periods_in_a_year_fixed:
@@ -1439,7 +1439,7 @@ class _CommonModel(BaseModel):
 
         dddf = (
             self.tsdf.loc[cast(int, earlier) : cast(int, later)]
-            .pct_change(fill_method=cast(str, None))
+            .pct_change()
             .sub(min_accepted_return / time_factor)
         )
 
@@ -1542,7 +1542,7 @@ class _CommonModel(BaseModel):
         )
         result: NDArray[float64] = skew(
             a=self.tsdf.loc[cast(int, earlier) : cast(int, later)]
-            .pct_change(fill_method=cast(str, None))
+            .pct_change()
             .to_numpy(),
             bias=True,
             nan_policy="omit",
@@ -1589,9 +1589,7 @@ class _CommonModel(BaseModel):
             to_dt=to_date,
         )
         result: NDArray[float64] = kurtosis(
-            self.tsdf.loc[cast(int, earlier) : cast(int, later)].pct_change(
-                fill_method=cast(str, None),
-            ),
+            self.tsdf.loc[cast(int, earlier) : cast(int, later)].pct_change(),
             fisher=True,
             bias=True,
             nan_policy="omit",
@@ -1687,19 +1685,13 @@ class _CommonModel(BaseModel):
         )
         pos = (
             self.tsdf.loc[cast(int, earlier) : cast(int, later)]
-            .pct_change(fill_method=cast(str, None))[1:][
-                self.tsdf.loc[cast(int, earlier) : cast(int, later)].pct_change(
-                    fill_method=cast(str, None),
-                )[1:]
+            .pct_change()[1:][
+                self.tsdf.loc[cast(int, earlier) : cast(int, later)].pct_change()[1:]
                 > zero
             ]
             .count()
         )
-        tot = (
-            self.tsdf.loc[cast(int, earlier) : cast(int, later)]
-            .pct_change(fill_method=cast(str, None))[1:]
-            .count()
-        )
+        tot = self.tsdf.loc[cast(int, earlier) : cast(int, later)].pct_change().count()
         share = pos / tot
         if self.tsdf.shape[1] == 1:
             return float(share.iloc[0])
@@ -1875,9 +1867,7 @@ class _CommonModel(BaseModel):
             from_dt=from_date,
             to_dt=to_date,
         )
-        retdf = self.tsdf.loc[cast(int, earlier) : cast(int, later)].pct_change(
-            fill_method=cast(str, None),
-        )
+        retdf = self.tsdf.loc[cast(int, earlier) : cast(int, later)].pct_change()
         pos = retdf[retdf > min_accepted_return].sub(min_accepted_return).sum()
         neg = retdf[retdf < min_accepted_return].sub(min_accepted_return).sum()
         ratio = pos / -neg
@@ -1965,7 +1955,7 @@ class _CommonModel(BaseModel):
             period = "-".join([str(year), str(month).zfill(2)])
         vrdf = self.tsdf.copy()
         vrdf.index = DatetimeIndex(vrdf.index)
-        resultdf = DataFrame(vrdf.pct_change(fill_method=None))  # type: ignore[arg-type]
+        resultdf = DataFrame(vrdf.pct_change())
         result = resultdf.loc[period] + 1
         cal_period = result.cumprod(axis="index").iloc[-1] - 1
         if self.tsdf.shape[1] == 1:
@@ -2017,7 +2007,7 @@ class _CommonModel(BaseModel):
         )
         result = (
             self.tsdf.loc[cast(int, earlier) : cast(int, later)]
-            .pct_change(fill_method=cast(str, None))
+            .pct_change()
             .quantile(1 - level, interpolation=interpolation)
         )
 
@@ -2065,7 +2055,7 @@ class _CommonModel(BaseModel):
         )
         result = (
             self.tsdf.loc[cast(int, earlier) : cast(int, later)]
-            .pct_change(fill_method=cast(str, None))
+            .pct_change()
             .rolling(observations, min_periods=observations)
             .sum()
             .min()
@@ -2111,9 +2101,7 @@ class _CommonModel(BaseModel):
             from_dt=from_date,
             to_dt=to_date,
         )
-        zscframe = self.tsdf.loc[cast(int, earlier) : cast(int, later)].pct_change(
-            fill_method=cast(str, None),
-        )
+        zscframe = self.tsdf.loc[cast(int, earlier) : cast(int, later)].pct_change()
         result = (zscframe.iloc[-1] - zscframe.mean()) / zscframe.std()
 
         if self.tsdf.shape[1] == 1:
@@ -2182,7 +2170,7 @@ class _CommonModel(BaseModel):
         ret_label = cast(tuple[str], self.tsdf.iloc[:, column].name)[0]
         retseries = (
             self.tsdf.iloc[:, column]
-            .pct_change(fill_method=cast(str, None))
+            .pct_change()
             .rolling(observations, min_periods=observations)
             .sum()
         )
@@ -2259,7 +2247,7 @@ class _CommonModel(BaseModel):
         else:
             time_factor = self.periods_in_a_year
         vol_label = cast(tuple[str, ValueType], self.tsdf.iloc[:, column].name)[0]
-        dframe = self.tsdf.iloc[:, column].pct_change(fill_method=cast(str, None))
+        dframe = self.tsdf.iloc[:, column].pct_change()
         volseries = dframe.rolling(
             observations,
             min_periods=observations,