+++ openseries/simulation.py
@@ -122,10 +122,7 @@ class ReturnSimulation(BaseModel):
         """
         return cast(
             float,
-            (
-                self.results.pct_change(fill_method=cast(str, None)).mean()
-                * self.trading_days_in_year
-            ).iloc[0],
+            (self.results.pct_change().mean() * self.trading_days_in_year).iloc[0],
         )
 
     @property
@@ -140,10 +137,9 @@ class ReturnSimulation(BaseModel):
         """
         return cast(
             float,
-            (
-                self.results.pct_change(fill_method=cast(str, None)).std()
-                * sqrt(self.trading_days_in_year)
-            ).iloc[0],
+            (self.results.pct_change().std() * sqrt(self.trading_days_in_year)).iloc[
+                0
+            ],
         )
 
     @classmethod