+++ openseries/types.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 import datetime as dt
 from enum import Enum
+from pprint import pformat
 from typing import Annotated, ClassVar, Literal, Union
 
 from numpy import datetime64
@@ -225,8 +226,8 @@ LiteralFrameProps = Literal[
 ]
 
 
-class OpenTimeSeriesPropertiesList(list[str]):
-    """Allowed property arguments for the OpenTimeSeries class."""
+class PropertiesList(list[str]):
+    """Base class for allowed property arguments definition."""
 
     allowed_strings: ClassVar[set[str]] = {
         "value_ret",
@@ -246,9 +247,38 @@ class OpenTimeSeriesPropertiesList(list[str]):
         "vol_from_var",
         "worst",
         "worst_month",
-        "max_drawdown_cal_year",
         "max_drawdown",
         "max_drawdown_date",
+        "max_drawdown_cal_year",
+    }
+
+    def _validate(self: Self) -> None:
+        """Validate the string input of the all_properties method."""
+        seen = set()
+        invalids = set()
+        duplicates = set()
+        msg = ""
+        for item in self:
+            if item not in self.allowed_strings:
+                invalids.add(item)
+            if item in seen:
+                duplicates.add(item)
+            seen.add(item)
+        if len(invalids) != 0:
+            msg += (
+                f"Invalid string(s): {list(invalids)}.\nAllowed strings are:"
+                f"\n{pformat(self.allowed_strings)}\n"
+            )
+        if len(duplicates) != 0:
+            msg += f"Duplicate string(s): {list(duplicates)}."
+        if len(msg) != 0:
+            raise ValueError(msg)
+
+
+class OpenTimeSeriesPropertiesList(PropertiesList):
+    """Allowed property arguments for the OpenTimeSeries class."""
+
+    allowed_strings: ClassVar[set[str]] = PropertiesList.allowed_strings | {
         "first_idx",
         "last_idx",
         "length",
@@ -265,44 +295,11 @@ class OpenTimeSeriesPropertiesList(list[str]):
         super().__init__(args)
         self._validate()
 
-    def _validate(self: Self) -> None:
-        seen = set()
-        for item in self:
-            if item not in self.allowed_strings:
-                msg = (
-                    f"Invalid string: {item}. Allowed strings: {self.allowed_strings}"
-                )
-                raise ValueError(msg)
-            if item in seen:
-                msg = f"Duplicate string: {item}"
-                raise ValueError(msg)
-            seen.add(item)
 
-
-class OpenFramePropertiesList(list[str]):
+class OpenFramePropertiesList(PropertiesList):
     """Allowed property arguments for the OpenFrame class."""
 
-    allowed_strings: ClassVar[set[str]] = {
-        "value_ret",
-        "geo_ret",
-        "arithmetic_ret",
-        "vol",
-        "downside_deviation",
-        "ret_vol_ratio",
-        "sortino_ratio",
-        "omega_ratio",
-        "z_score",
-        "skew",
-        "kurtosis",
-        "positive_share",
-        "var_down",
-        "cvar_down",
-        "vol_from_var",
-        "worst",
-        "worst_month",
-        "max_drawdown",
-        "max_drawdown_date",
-        "max_drawdown_cal_year",
+    allowed_strings: ClassVar[set[str]] = PropertiesList.allowed_strings | {
         "first_indices",
         "last_indices",
         "lengths_of_items",
@@ -313,19 +310,6 @@ class OpenFramePropertiesList(list[str]):
         """Property arguments for the OpenFrame class."""
         super().__init__(args)
         self._validate()
-
-    def _validate(self: Self) -> None:
-        seen = set()
-        for item in self:
-            if item not in self.allowed_strings:
-                msg = (
-                    f"Invalid string: {item}. Allowed strings: {self.allowed_strings}"
-                )
-                raise ValueError(msg)
-            if item in seen:
-                msg = f"Duplicate string: {item}"
-                raise ValueError(msg)
-            seen.add(item)
 
 
 class ValueType(str, Enum):