+++ tests/test_types.py
@@ -25,24 +25,35 @@ class TestTypes:
         if not isinstance(lst, OpenTimeSeriesPropertiesList):
             raise TypeError(msg)
 
-        with pytest.raises(
-            expected_exception=ValueError,
-            match="Invalid string: boo",
-        ):
+        msg2 = "OpenTimeSeriesPropertiesList._validate not raising expected Exception"
+        with pytest.raises(expected_exception=ValueError) as exc:
             OpenTimeSeriesPropertiesList(
                 *cast(LiteralSeriesProps, ["z_score", "boo", "positive_share"]),
             )
+        if "Invalid string(s): ['boo']." not in str(exc.value):
+            raise ValueError(msg2)
 
-        with pytest.raises(
-            expected_exception=ValueError,
-            match="Duplicate string: skew",
-        ):
+        with pytest.raises(expected_exception=ValueError) as excc:
             OpenTimeSeriesPropertiesList(
                 *cast(
                     LiteralSeriesProps,
                     ["z_score", "skew", "skew", "positive_share"],
                 ),
             )
+        if "Duplicate string(s): ['skew']." not in str(excc.value):
+            raise ValueError(msg2)
+
+        with pytest.raises(expected_exception=ValueError) as exccc:
+            OpenTimeSeriesPropertiesList(
+                *cast(
+                    LiteralSeriesProps,
+                    ["z_score", "skew", "skew", "boo"],
+                ),
+            )
+        if "Duplicate string(s): ['skew']." not in str(
+            exccc.value,
+        ) and "Invalid string(s): ['boo']." not in str(exccc.value):
+            raise ValueError(msg2)
 
     def test_openframeproplist_validate(self: TestTypes) -> None:
         """Test that the OpenFrame property input is correctly checked."""
@@ -52,18 +63,32 @@ class TestTypes:
         if not isinstance(lst, OpenFramePropertiesList):
             raise TypeError(msg)
 
-        with pytest.raises(expected_exception=ValueError, match="Invalid string: boo"):
+        msg2 = "OpenFramePropertiesList._validate not raising expected Exception"
+        with pytest.raises(expected_exception=ValueError) as exc:
             OpenFramePropertiesList(
                 *cast(LiteralFrameProps, ["z_score", "boo", "positive_share"]),
             )
+        if "Invalid string(s): ['boo']." not in str(exc.value):
+            raise ValueError(msg2)
 
-        with pytest.raises(
-            expected_exception=ValueError,
-            match="Duplicate string: skew",
-        ):
+        with pytest.raises(expected_exception=ValueError) as excc:
             OpenFramePropertiesList(
                 *cast(
                     LiteralFrameProps,
                     ["z_score", "skew", "skew", "positive_share"],
                 ),
             )
+        if "Duplicate string(s): ['skew']." not in str(excc.value):
+            raise ValueError(msg2)
+
+        with pytest.raises(expected_exception=ValueError) as exccc:
+            OpenFramePropertiesList(
+                *cast(
+                    LiteralFrameProps,
+                    ["z_score", "skew", "skew", "boo"],
+                ),
+            )
+        if "Duplicate string(s): ['skew']." not in str(
+            exccc.value,
+        ) and "Invalid string(s): ['boo']." not in str(exccc.value):
+            raise ValueError(msg2)